refactor: simplify ChatContent component and enhance page state handling
- Removed unused imports and commented-out code in ChatContent.
- Replaced the manual rendering of the disabled state with a new Disabled component for better readability.
- Updated Input component to include pageState in the condition for message submission.
- Adjusted ChatContextProvider to correctly determine page state based on agent settings and loading states.
- Enhanced useChatParams hook to prevent unnecessary refetching and caching.
