+++ packages/app/src/react/features/ai-chat/components/index.ts
@@ -2,8 +2,8 @@ export * from './chat';
 export * from './chat-content';
 export * from './common';
 export * from './container';
+export * from './disabled';
 export * from './footer';
 export * from './header';
 export * from './loading-skeleton';
 export * from './skeleton';
-
